DAOS-5180 test: Include HDF5 in soak
Test-tag-hw-large:pr,hw,large soak_smoke
