+++ src/tests/ftest/soak/soak_stress_2h.yaml
@@ -136,6 +136,7 @@ ior_stress:
         - '64k'
     dfs_oclass:
         - "SX"
+    dfs_destroy: True
 fio_stress:
   names:
     - global
